Trying new pathfinding
Going to try to use PathNavigator to actually move on the path from NavMesh.CalculatePath. Want to put NavMeshAgents on the characters so that the calculated paths avoid them, and set the layer mask to nothing so hopefully they won't control the characters at all, just mark them as obstacles. NavMeshObstacle didn't seem to work, could try that again too.