Working on transparent circle shader.
Getting better, but still problems. The main material that is using is blank white now. How to keep the colors? Need to adjust the sphere colliders radius to fit with the transparent circle too. Also need to keep track of which materials just stopped being in the way, so hopefully things behind the player don't get shadered too.